requirements.txt
# requirements.txt
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
scipy>=1.11.0
numpy>=1.24.0